+++ sites/next.skeleton.dev/src/components/astro/ChipsBar.astro
