+++ sites/next.skeleton.dev/src/components/astro/Table.astro
