+++ sites/next.skeleton.dev/src/components/mdx/Heading2.astro
