+++ sites/next.skeleton.dev/src/components/mdx/Heading6.astro
