+++ sites/next.skeleton.dev/src/components/mdx/HorzRule.astro
