+++ sites/next.skeleton.dev/src/components/mdx/Table.astro
