+++ sites/next.skeleton.dev/src/components/react/FrameworkTabs.tsx
