+++ sites/next.skeleton.dev/src/content/components/react/accordions.mdx
@@ -1,11 +1,11 @@
 ---
-layout: '@lib/layouts/LayoutDoc.astro'
+layout: '@layouts/LayoutDoc.astro'
 ---
 
 export const components = componentSet;
 
-import { Page as Example } from "@lib/previews/components/accordions/Example.tsx";
-import ExampleRaw from "@lib/previews/components/accordions/Example.tsx?raw";
+import { Page as Example } from "@examples/components/accordions/Example.tsx";
+import ExampleRaw from "@examples/components/accordions/Example.tsx?raw";
 
 import * as Common from '@content/components/common/accordions.mdx';
