+++ sites/next.skeleton.dev/src/content/components/svelte/accordions.mdx
@@ -1,11 +1,11 @@
 ---
-layout: '@lib/layouts/LayoutDoc.astro'
+layout: '@layouts/LayoutDoc.astro'
 ---
 
 export const components = componentSet;
 
-import Example from "@lib/previews/components/accordions/Example.svelte";
-import ExampleRaw from "@lib/previews/components/accordions/Example.svelte?raw";
+import Example from "@examples/components/accordions/Example.svelte";
+import ExampleRaw from "@examples/components/accordions/Example.svelte?raw";
 
 import * as Common from '@content/components/common/accordions.mdx';
