+++ sites/next.skeleton.dev/src/content/components/svelte/app-bars.mdx
@@ -1,5 +1,5 @@
 ---
-layout: '@lib/layouts/LayoutDoc.astro'
+layout: '@layouts/LayoutDoc.astro'
 ---
 
 export const components = componentSet
