+++ sites/next.skeleton.dev/src/content/design/colors.mdx
@@ -1,5 +1,5 @@
 ---
-layout: '@lib/layouts/LayoutDoc.astro'
+layout: '@layouts/LayoutDoc.astro'
 title: Colors
 description: The Skeleton color system.
 srcCore: '/plugin/extends/colors.ts'
@@ -8,8 +8,8 @@ showDocsUrl: true
 
 export const components = componentSet
 
-import Swatches from '@lib/previews/design/colors/Swatches.astro';
-import Pairings from '@lib/previews/design/colors/Pairings.astro';
+import Swatches from '@examples/design/colors/Swatches.astro';
+import Pairings from '@examples/design/colors/Pairings.astro';
 
 ## Palette
