+++ sites/next.skeleton.dev/src/content/design/presets.mdx
@@ -1,5 +1,5 @@
 ---
-layout: '@lib/layouts/LayoutDoc.astro'
+layout: '@layouts/LayoutDoc.astro'
 title: Presets
 description: Provides a robust preset combination of styles for use with buttons, cards, badges, and more.
 srcCore: '/plugin/utilities/presets.ts'
