+++ sites/next.skeleton.dev/src/content/design/spacing.mdx
@@ -1,5 +1,5 @@
 ---
-layout: '@lib/layouts/LayoutDoc.astro'
+layout: '@layouts/LayoutDoc.astro'
 title: Spacing
 description: The Skeleton spacing system.
 srcCore: '/plugin/extends/spacing.ts'
@@ -8,7 +8,7 @@ showDocsUrl: true
 
 export const components = componentSet
 
-import TableSpace from "@lib/previews/design/spacing/TableSpace.astro";
+import TableSpace from "@examples/design/spacing/TableSpace.astro";
 
 ## Supports
