+++ sites/next.skeleton.dev/src/content/design/typography.mdx
@@ -1,5 +1,5 @@
 ---
-layout: '@lib/layouts/LayoutDoc.astro'
+layout: '@layouts/LayoutDoc.astro'
 title: Typography
 description: The Skeleton typography system.
 srcCore: '/plugin/components/typography.css'
