+++ sites/next.skeleton.dev/src/content/docs/introduction.mdx
@@ -1,5 +1,5 @@
 ---
-layout: '@lib/layouts/LayoutDoc.astro'
+layout: '@layouts/LayoutDoc.astro'
 title: Introduction
 description: (work in progress)
 ---
