+++ sites/next.skeleton.dev/src/content/resources/cookbook.mdx
@@ -1,12 +1,12 @@
 ---
-layout: '@lib/layouts/LayoutDoc.astro'
+layout: '@layouts/LayoutDoc.astro'
 title: Cookbook
 description: A collection of recipes for crafting complex UI that combines Skeleton and Tailwind primitives.
 ---
 
 export const components = componentSet
 
-import NavGrid from "@lib/previews/resources/cookbook/NavGrid.astro";
+import NavGrid from "@examples/resources/cookbook/NavGrid.astro";
 
 {/* --- */}
