+++ sites/next.skeleton.dev/src/content/resources/cookbook/alerts.mdx
@@ -1,5 +1,5 @@
 ---
-layout: '@lib/layouts/LayoutDoc.astro'
+layout: '@layouts/LayoutDoc.astro'
 title: Alerts
 description: Displays customizable alerts to attract attention and provide critical actions.
 showDocsUrl: true
