+++ sites/next.skeleton.dev/src/content/resources/cookbook/breadcrumbs.mdx
@@ -1,5 +1,5 @@
 ---
-layout: '@lib/layouts/LayoutDoc.astro'
+layout: '@layouts/LayoutDoc.astro'
 title: Breadcrumbs
 description: Displays the current navigation hierarchy.
 showDocsUrl: true
