+++ sites/next.skeleton.dev/src/content/resources/cookbook/logo-clouds.mdx
@@ -1,5 +1,5 @@
 ---
-layout: '@lib/layouts/LayoutDoc.astro'
+layout: '@layouts/LayoutDoc.astro'
 title: Logo Clouds
 description: Provides a grid for presenting a set of logos, brands, or sponsors.
 showDocsUrl: true
