+++ sites/next.skeleton.dev/src/content/resources/markdown.mdx
@@ -1,5 +1,5 @@
 ---
-layout: '@lib/layouts/LayoutDoc.astro'
+layout: '@layouts/LayoutDoc.astro'
 title: Markdown
 description: An MDX kitchen sink to test parsing and styles.
 ---
