+++ sites/next.skeleton.dev/src/content/tailwind/badges.mdx
@@ -1,5 +1,5 @@
 ---
-layout: '@lib/layouts/LayoutDoc.astro'
+layout: '@layouts/LayoutDoc.astro'
 title: Badges
 description: Provides a robust set of non-interactive badge styles.
 srcCore: '/plugin/components/badges.css'
