+++ sites/next.skeleton.dev/src/content/tailwind/buttons.mdx
@@ -1,5 +1,5 @@
 ---
-layout: '@lib/layouts/LayoutDoc.astro'
+layout: '@layouts/LayoutDoc.astro'
 title: Buttons
 description: Provide a variety of button, including customizable sizes and types.
 srcCore: '/plugin/components/buttons.css'
