+++ sites/next.skeleton.dev/src/content/tailwind/cards.mdx
@@ -1,5 +1,5 @@
 ---
-layout: '@lib/layouts/LayoutDoc.astro'
+layout: '@layouts/LayoutDoc.astro'
 title: Cards
 description: Provides container elements that wrap and separate content.
 srcCore: '/plugin/components/cards.css'
