+++ sites/next.skeleton.dev/src/content/tailwind/chips.mdx
@@ -1,5 +1,5 @@
 ---
-layout: '@lib/layouts/LayoutDoc.astro'
+layout: '@layouts/LayoutDoc.astro'
 title: Chips
 description: Provides a robust set of interactive chip styles.
 srcCore: '/plugin/components/chips.css'
@@ -8,7 +8,7 @@ showDocsUrl: true
 
 export const components = componentSet
 
-import Selection from "@lib/previews/tailwind/chips/Selection.svelte";
+import Selection from "@examples/tailwind/chips/Selection.svelte";
 
 <div class="docs-card-outlined-centered p-8">
     <div class="flex items-center gap-4">
