+++ sites/next.skeleton.dev/src/content/tailwind/dividers.mdx
@@ -1,5 +1,5 @@
 ---
-layout: '@lib/layouts/LayoutDoc.astro'
+layout: '@layouts/LayoutDoc.astro'
 title: Dividers
 description: Horizontal and vertical rule styling.
 srcCore: '/plugin/components/dividers.css'
