+++ sites/next.skeleton.dev/src/content/tailwind/forms.mdx
@@ -1,5 +1,5 @@
 ---
-layout: '@lib/layouts/LayoutDoc.astro'
+layout: '@layouts/LayoutDoc.astro'
 title: Forms & Inputs
 description: Various form and input styles.
 srcCore: '/plugin/components/forms.css'
