+++ sites/next.skeleton.dev/src/content/tailwind/placeholders.mdx
@@ -1,5 +1,5 @@
 ---
-layout: '@lib/layouts/LayoutDoc.astro'
+layout: '@layouts/LayoutDoc.astro'
 title: Placeholders
 description: Provides "skeleton" placeholders that can display while content loads.
 srcCore: '/plugin/components/placeholders.css'
