+++ sites/next.skeleton.dev/src/content/tailwind/tables.mdx
@@ -1,5 +1,5 @@
 ---
-layout: '@lib/layouts/LayoutDoc.astro'
+layout: '@layouts/LayoutDoc.astro'
 title: Tables
 description: Provides a set of styles for native HTML table elements.
 srcCore: '/plugin/components/tables.css'
@@ -8,7 +8,7 @@ showDocsUrl: true
 
 export const components = componentSet
 
-import Example from "@lib/previews/tailwind/tables/Example.astro";
+import Example from "@examples/tailwind/tables/Example.astro";
 
 <div class="docs-card-outlined-centered p-8">
     <Example />
