+++ sites/next.skeleton.dev/src/examples/components/accordions/Example.svelte
