+++ sites/next.skeleton.dev/src/examples/design/colors/Pairings.astro
