+++ sites/next.skeleton.dev/src/examples/design/colors/Swatches.astro
