+++ sites/next.skeleton.dev/src/examples/design/spacing/TableSpace.astro
