+++ sites/next.skeleton.dev/src/examples/resources/cookbook/NavGrid.astro
