+++ sites/next.skeleton.dev/src/examples/tailwind/tables/Example.astro
