+++ sites/next.skeleton.dev/src/layouts/LayoutDoc.astro
@@ -1,14 +1,14 @@
 ---
 import { Icon } from 'astro-icon/components'
 
-import LayoutRoot from '@lib/layouts/LayoutRoot.astro'
+import LayoutRoot from '@layouts/LayoutRoot.astro'
 
-import Header from '@lib/components/astro/Header.astro'
-import ChipsBar from '@lib/components/astro/ChipsBar.astro'
-import Navigation from '@lib/components/astro/Navigation.astro'
-import Breadcrumbs from '@lib/components/astro/Breadcrumbs.astro'
-import FrameworkPicker from '@lib/components/astro/FrameworkPicker.astro'
-import TableOfContents from '@lib/components/astro/TableOfContents.astro'
+import Header from '@components/astro/Header.astro'
+import ChipsBar from '@components/astro/ChipsBar.astro'
+import Navigation from '@components/astro/Navigation.astro'
+import Breadcrumbs from '@components/astro/Breadcrumbs.astro'
+import FrameworkPicker from '@components/astro/FrameworkPicker.astro'
+import TableOfContents from '@components/astro/TableOfContents.astro'
 
 interface PageHeadings {
     depth: number
