+++ sites/next.skeleton.dev/src/pages/index.astro
@@ -1,8 +1,8 @@
 ---
 import { Icon } from 'astro-icon/components'
 
-import LayoutRoot from '@lib/layouts/LayoutRoot.astro'
-import Header from '@lib/components/astro/Header.astro'
+import LayoutRoot from '@layouts/LayoutRoot.astro'
+import Header from '@components/astro/Header.astro'
 
 const layoutProps = {
     classList: 'space-y-10',
